Выносит хуки в отдельные модули
