+++ src/components/video-player/video-player.jsx
@@ -1,33 +1,6 @@
-import React, {createRef, useEffect} from "react";
+import React, {createRef} from "react";
 import PropTypes from "prop-types";
-
-const useVideoPlayer = (videoRef, onCurrentTimeChange, resetAfterPause, isPlaying) => {
-  useEffect(() => {
-    const video = videoRef.current;
-
-    video.ontimeupdate = () => {
-      const currentTime = Math.floor(video.currentTime);
-
-      if (onCurrentTimeChange) {
-        onCurrentTimeChange(currentTime);
-      }
-    };
-  }, []);
-
-  useEffect(() => {
-    const video = videoRef.current;
-
-    if (isPlaying) {
-      video.play();
-    } else {
-      video.pause();
-
-      if (resetAfterPause) {
-        video.load();
-      }
-    }
-  }, [isPlaying, resetAfterPause]);
-};
+import {useVideoPlayer} from "../../hooks/use-video-player/use-video-player";
 
 const VideoPlayer = (props) => {
   const {
