+++ src/hooks/use-film-list-loader/use-film-list-loader.jsx
@@ -1,0 +1,7 @@
+import {useEffect} from "react";
+
+export const useFilmListLoader = (loadFilmList) => {
+  useEffect(() => {
+    loadFilmList();
+  }, []);
+};
