+++ src/hooks/use-promo-film-loader/use-promo-film-loader.jsx
@@ -1,0 +1,7 @@
+import {useEffect} from "react";
+
+export const usePromoFilmLoader = (loadPromoFilm) => {
+  useEffect(() => {
+    loadPromoFilm();
+  }, []);
+};
