+++ src/hooks/use-reviews-data/use-reviews-data.jsx
@@ -1,0 +1,7 @@
+import {useEffect} from "react";
+
+export const useReviewsData = (filmId, loadReviews) => {
+  useEffect(() => {
+    loadReviews(filmId);
+  }, [filmId]);
+};
